# app/Models/Attribute.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Relations\HasMany;

class Attribute extends Model
{
    protected $fillable = ['name', 'type'];

    public function attributeValues(): HasMany
    {
        return $this->hasMany(AttributeValue::class);
    }
}