# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/attributes": {
            "get": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Get all attributes",
                "operationId": "e48983aaee64bf1da038f1378aff0c81",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Attribute"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Create a new attribute",
                "operationId": "48e2284fe9ed7ecdc495aa7191e74084",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AttributeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attribute created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attribute"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attributes/{attribute}": {
            "get": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Get a single attribute by ID",
                "operationId": "a7e093c764f78918851a4bb3220f27b4",
                "parameters": [
                    {
                        "name": "attribute",
                        "in": "path",
                        "description": "Attribute ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attribute"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Attribute not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Update an existing attribute",
                "operationId": "a7fef12bab86c4124d12cfb2a9f57452",
                "parameters": [
                    {
                        "name": "attribute",
                        "in": "path",
                        "description": "Attribute ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AttributeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attribute updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attribute"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Delete an attribute",
                "operationId": "ac4021e09ece6c1e01a5746946d8bebf",
                "parameters": [
                    {
                        "name": "attribute",
                        "in": "path",
                        "description": "Attribute ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attribute deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Attribute deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attributeValues": {
            "get": {
                "tags": [
                    "Attribute Values"
                ],
                "summary": "Get all attributes values",
                "operationId": "8acdc4d4b1d54b4a64d45701a7886de6",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AttributeValues"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attributeValues/{attributeValue}": {
            "get": {
                "tags": [
                    "Attribute Values"
                ],
                "summary": "Get a single attribute value by ID",
                "operationId": "98f5fbdac301999b8a0ca23f53d003c6",
                "parameters": [
                    {
                        "name": "attributeValue",
                        "in": "path",
                        "description": "Attribute Value ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttributeValues"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Attribute Value not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Attribute Values"
                ],
                "summary": "Delete an attribute value",
                "operationId": "c8fc014c69b036bd58dd2ba8cb31409c",
                "parameters": [
                    {
                        "name": "attributeValue",
                        "in": "path",
                        "description": "Attribute Value ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attribute Value deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AttributeRequest": {
                "title": "Attribute Request",
                "description": "Schema for creating an attribute",
                "required": [
                    "name",
                    "type"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "department",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "text",
                            "number",
                            "date",
                            "select"
                        ],
                        "example": "text",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Attribute": {
                "title": "Attribute",
                "description": "Attribute model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "department"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "text",
                            "number",
                            "date",
                            "select"
                        ],
                        "example": "text"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "AttributeValues": {
                "title": "Attribute Values",
                "description": "Attribute Values model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "attribute_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "entity_id": {
                        "type": "integer",
                        "example": "product_id"
                    },
                    "value": {
                        "type": "string",
                        "example": "IT"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}